moved types and output layout function out of file
